update autocomplete fetchers
I fixed the project SORT issue and while I was at it corrected some of the other fetchers. Take a look and let me know if you have questions on why I did what I did. For the standards and the courses, you don't need the UNIQUE(FLATTEN()). It's only when we're pulling an array of arrays.